:fire: refactor: 删除命名空间
